Refactor EditArtist and NewArtist to use shared form
